+++ src/pages/Order/CartTable.scss
@@ -29,7 +29,7 @@
       }
       td {
         border: none;
-        padding: 10px;
+        padding: 0.8rem;
       }
 
       .pricebox {
