+++ src/pages/Order/Footer.scss
@@ -1,0 +1,26 @@
+.footer {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  width: 80rem;
+  border-top: 1px solid black;
+  padding: 1rem 0;
+  margin-top: 1.25rem;
+
+  .total {
+    display: flex;
+    align-items: center;
+
+    p {
+      margin-right: 20px;
+    }
+  }
+
+  button {
+    border: none;
+    background-color: inherit;
+    font-weight: bold;
+    font-size: 16px;
+    cursor: pointer;
+  }
+}
