+++ src/pages/Order/OrderInfo.scss
@@ -9,7 +9,7 @@
   th {
     text-align: start;
     padding-right: 4rem;
-    padding-bottom: 0.5rem;
+    padding-bottom: 1rem;
     font-weight: 400;
   }
 
@@ -61,6 +61,7 @@
     }
   }
 }
+
 .messageBox {
   outline: none;
   border: none;
