+++ src/pages/Order/OrderPage.scss
@@ -3,10 +3,11 @@
   flex-direction: column;
   justify-content: center;
   align-items: center;
-  padding: 16rem 0;
+  padding: 10rem 0;
 
   h1 {
     text-align: center;
     font-size: 25px;
+    margin-bottom: 1rem;
   }
 }
