+++ src/pages/Order/Payment.scss
@@ -78,28 +78,3 @@
     outline: none;
   }
 }
-
-.footer {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  border-top: 1px solid black;
-  margin-top: 1.25rem;
-
-  .total {
-    display: flex;
-    align-items: center;
-
-    p {
-      margin-right: 20px;
-    }
-  }
-
-  button {
-    border: none;
-    background-color: inherit;
-    font-weight: bold;
-    font-size: 15px;
-    cursor: pointer;
-  }
-}
